created product controller

+++ database/migrations/2022_07_11_111445_create_product_price.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->increments('ID')->unsigned();
             $table->integer('company_id')->unsigned()->nullable();
             $table->integer('product_id')->unsigned()->nullable();
-            $table->integer('price')->nullable();
+            $table->bigInteger('price')->nullable();
             $table->timestamp('published_at')->nullable();
 
             // $table->foreign('company_id')->references('ID')
